requirements.txt
streamlit>=1.25.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.173
ollama>=0.1.0
